- Reorganizing what used to be in src/manifest.ts into a separate directory - Relocating manifest related types from src/types.ts. - Creating an ISlackManifestShared to contain types and features that are common to both Run on Slack and Remote app scenarios - Relocated all ManifestSchema types into its own file. Since these pertain directly to internal types it made sense to keep them in their own little corner. - Omitted runOnSlack from Manifest function arguments type
