--- .github/workflows/npm-publish.yml
@@ -1,46 +1,0 @@
-# This workflow will run tests using node and then publish a package to GitHub Packages when a release is created
-# For more information see: https://help.github.com/actions/language-and-framework-guides/publishing-nodejs-packages
-
-name: Build & Deploy to NPM
-
-on:
-  push:    
-    tags:        
-      - '*'
-    branches: [ main ]
-
-jobs:
-
-  build:
-    runs-on: macos-latest
-    
-    steps:
-      - name: Actions checkout
-        uses: actions/checkout@v3
-      
-      - name: Setup node
-        uses: actions/setup-node@v3
-        with:
-          node-version: 16
-          registry-url: https://registry.npmjs.org/
-          
-      - name: Setup Deno
-        uses: denoland/setup-deno@v1
-        with:
-          deno-version: v1.x
-          
-      - name: Run build_npm.ts
-        run: deno run -A build_npm.ts 0.1.0
-         
-      - name: Publish to NPM    
-        run: cd npm && npm publish
-        env:
-          NODE_AUTH_TOKEN: ${{secrets.npm_token}}
-        
-         
-    
-    
-
-      
-
-
