+++ src/datastore/types.ts
@@ -1,6 +1,6 @@
 import { ICustomType } from "../types/types.ts";
-import { ManifestDatastoreSchema } from "../types.ts";
-import { SlackManifest } from "../manifest.ts";
+import { ManifestDatastoreSchema } from "../manifest/manifest_schema.ts";
+import { SlackManifest } from "../manifest/mod.ts";
 
 export type SlackDatastoreAttribute = {
   // supports custom types, primitive types, inline objects and lists
