+++ src/dev_deps.ts
@@ -3,3 +3,7 @@ export {
   assertExists,
   assertStrictEquals,
 } from "https://deno.land/std@0.134.0/testing/asserts.ts";
+
+export { assert } from "https://deno.land/x/conditional_type_checks@1.0.6/mod.ts";
+
+export type { IsExact } from "https://deno.land/x/conditional_type_checks@1.0.6/mod.ts";
