+++ src/mod.ts
@@ -1,4 +1,10 @@
-export { Manifest } from "./manifest.ts";
+export { Manifest, SlackManifest } from "./manifest/mod.ts";
+export type {
+  ISlackManifestRemote,
+  ISlackManifestRunOnSlack,
+  SlackManifestType,
+} from "./manifest/types.ts";
+export type { ManifestSchema } from "./manifest/manifest_schema.ts";
 export { DefineFunction } from "./functions/mod.ts";
 export { DefineWorkflow } from "./workflows/mod.ts";
 export { DefineType } from "./types/mod.ts";
