+++ src/types/mod.ts
@@ -1,5 +1,5 @@
-import { SlackManifest } from "../manifest.ts";
-import { ManifestCustomTypeSchema } from "../types.ts";
+import { SlackManifest } from "../manifest/mod.ts";
+import { ManifestCustomTypeSchema } from "../manifest/manifest_schema.ts";
 import {
   CustomTypeDefinition,
   DefineTypeFunction,
