+++ src/workflows/types.ts
@@ -1,5 +1,5 @@
-import { SlackManifest } from "../manifest.ts";
-import type { ManifestWorkflowSchema } from "../types.ts";
+import { SlackManifest } from "../manifest/mod.ts";
+import type { ManifestWorkflowSchema } from "../manifest/manifest_schema.ts";
 import {
   ParameterPropertiesDefinition,
   ParameterSetDefinition,
