+++ src/workflows/workflow-step.ts
@@ -1,5 +1,8 @@
-import { SlackManifest } from "../manifest.ts";
-import { ManifestFunction, ManifestWorkflowStepSchema } from "../types.ts";
+import { SlackManifest } from "../manifest/mod.ts";
+import {
+  ManifestFunction,
+  ManifestWorkflowStepSchema,
+} from "../manifest/manifest_schema.ts";
 import { ISlackFunction } from "../functions/types.ts";
 import {
   CreateUntypedObjectParameterVariable,
